+++ 30_sources/SRC_Zupfnoter/src/config-form.rb
@@ -329,7 +329,7 @@ class ConfstackEditor
           Integer         => ['startpos', 'pack_method', 'p_repeat', 'p_begin', 'p_end', 'p_follow', 'PITCH_OFFSET'],
           OneLineString   => ['title', 'filenamepart', 'url', 'filebase'],
           MultiLineString => ['text'],
-          Boolean         => ['limit_a3', 'autopos', 'bottomup', 'show_border', 'nonflowrest', "show", "fill", "grid"],
+          Boolean         => ['limit_a3', 'autopos', 'bottomup', 'beams', 'show_border', 'nonflowrest', "show", "fill", "grid"],
           Float           => ['LINE_THIN', 'LINE_MEDIUM', 'LINE_THICK', 'pack_max_spreadfactor', 'pack_min_increment', 'minc_f', "X_SPACING", "X_OFFSET"],
           TupletShape     => ['shape'],
           TextStyle       => ['style'],