+++ 30_sources/CHANGES.md
@@ -43,6 +43,7 @@
 * support for more diatonic instruments (e.g. OKON harp) #196
 * support of okon table harp #196
 * support to write notes bottom up #196
+* support to write notes with beams #196
 * show notes in the current selection for harmonization work-in-progress #190
 
 # v 1.7.1